common update

+++ mybatis-common-mapper/src/main/java/com/dsy/dao/domain/salary/SocialCitydata.java
@@ -22,25 +22,25 @@ public class SocialCitydata {
      * 社保最高基数
      */
     @Column(name = "social_security_max_base")
-    private Integer socialSecurityMaxBase;
+    private BigDecimal socialSecurityMaxBase;
 
     /**
      * 社保最低基数
      */
     @Column(name = "social_security_min_base")
-    private Integer socialSecurityMinBase;
+    private BigDecimal socialSecurityMinBase;
 
     /**
      * 养老保险最高基数
      */
     @Column(name = "ylao_max_base")
-    private Integer ylaoMaxBase;
+    private BigDecimal ylaoMaxBase;
 
     /**
      * 养老保险最低基数
      */
     @Column(name = "ylao_min_base")
-    private Integer ylaoMinBase;
+    private BigDecimal ylaoMinBase;
 
     /**
      * 养老保险企业比例
@@ -58,13 +58,13 @@ public class SocialCitydata {
      * 医疗保险最高基数
      */
     @Column(name = "yliao_max_base")
-    private Integer yliaoMaxBase;
+    private BigDecimal yliaoMaxBase;
 
     /**
      * 医疗保险最低基数
      */
     @Column(name = "yliao_min_base")
-    private Integer yliaoMinBase;
+    private BigDecimal yliaoMinBase;
 
     /**
      * 医疗保险企业比例
@@ -82,13 +82,13 @@ public class SocialCitydata {
      * 失业保险最高基数
      */
     @Column(name = "sye_max_base")
-    private Integer syeMaxBase;
+    private BigDecimal syeMaxBase;
 
     /**
      * 失业保险最低基数
      */
     @Column(name = "sye_min_base")
-    private Integer syeMinBase;
+    private BigDecimal syeMinBase;
 
     /**
      * 失业保险企业比例
@@ -106,13 +106,13 @@ public class SocialCitydata {
      * 工伤保险最高基数
      */
     @Column(name = "gs_max_base")
-    private Integer gsMaxBase;
+    private BigDecimal gsMaxBase;
 
     /**
      * 工伤保险最低基数
      */
     @Column(name = "gs_min_base")
-    private Integer gsMinBase;
+    private BigDecimal gsMinBase;
 
     /**
      * 工伤保险企业比例
@@ -130,13 +130,13 @@ public class SocialCitydata {
      * 生育保险最高基数
      */
     @Column(name = "syu_max_base")
-    private Integer syuMaxBase;
+    private BigDecimal syuMaxBase;
 
     /**
      * 生育保险最低基数
      */
     @Column(name = "syu_min_base")
-    private Integer syuMinBase;
+    private BigDecimal syuMinBase;
 
     /**
      * 生育保险企业比例
@@ -154,13 +154,13 @@ public class SocialCitydata {
      * 住房公积金最高基数
      */
     @Column(name = "zf_max_base")
-    private Integer zfMaxBase;
+    private BigDecimal zfMaxBase;
 
     /**
      * 住房公积金最低基数
      */
     @Column(name = "zf_min_base")
-    private Integer zfMinBase;
+    private BigDecimal zfMinBase;
 
     /**
      * 住房公积金企业比例
@@ -220,7 +220,7 @@ public class SocialCitydata {
      *
      * @return social_security_max_base - 社保最高基数
      */
-    public Integer getSocialSecurityMaxBase() {
+    public BigDecimal getSocialSecurityMaxBase() {
         return socialSecurityMaxBase;
     }
 
@@ -229,7 +229,7 @@ public class SocialCitydata {
      *
      * @param socialSecurityMaxBase 社保最高基数
      */
-    public void setSocialSecurityMaxBase(Integer socialSecurityMaxBase) {
+    public void setSocialSecurityMaxBase(BigDecimal socialSecurityMaxBase) {
         this.socialSecurityMaxBase = socialSecurityMaxBase;
     }
 
@@ -238,7 +238,7 @@ public class SocialCitydata {
      *
      * @return social_security_min_base - 社保最低基数
      */
-    public Integer getSocialSecurityMinBase() {
+    public BigDecimal getSocialSecurityMinBase() {
         return socialSecurityMinBase;
     }
 
@@ -247,7 +247,7 @@ public class SocialCitydata {
      *
      * @param socialSecurityMinBase 社保最低基数
      */
-    public void setSocialSecurityMinBase(Integer socialSecurityMinBase) {
+    public void setSocialSecurityMinBase(BigDecimal socialSecurityMinBase) {
         this.socialSecurityMinBase = socialSecurityMinBase;
     }
 
@@ -256,7 +256,7 @@ public class SocialCitydata {
      *
      * @return ylao_max_base - 养老保险最高基数
      */
-    public Integer getYlaoMaxBase() {
+    public BigDecimal getYlaoMaxBase() {
         return ylaoMaxBase;
     }
 
@@ -265,7 +265,7 @@ public class SocialCitydata {
      *
      * @param ylaoMaxBase 养老保险最高基数
      */
-    public void setYlaoMaxBase(Integer ylaoMaxBase) {
+    public void setYlaoMaxBase(BigDecimal ylaoMaxBase) {
         this.ylaoMaxBase = ylaoMaxBase;
     }
 
@@ -274,7 +274,7 @@ public class SocialCitydata {
      *
      * @return ylao_min_base - 养老保险最低基数
      */
-    public Integer getYlaoMinBase() {
+    public BigDecimal getYlaoMinBase() {
         return ylaoMinBase;
     }
 
@@ -283,7 +283,7 @@ public class SocialCitydata {
      *
      * @param ylaoMinBase 养老保险最低基数
      */
-    public void setYlaoMinBase(Integer ylaoMinBase) {
+    public void setYlaoMinBase(BigDecimal ylaoMinBase) {
         this.ylaoMinBase = ylaoMinBase;
     }
 
@@ -328,7 +328,7 @@ public class SocialCitydata {
      *
      * @return yliao_max_base - 医疗保险最高基数
      */
-    public Integer getYliaoMaxBase() {
+    public BigDecimal getYliaoMaxBase() {
         return yliaoMaxBase;
     }
 
@@ -337,7 +337,7 @@ public class SocialCitydata {
      *
      * @param yliaoMaxBase 医疗保险最高基数
      */
-    public void setYliaoMaxBase(Integer yliaoMaxBase) {
+    public void setYliaoMaxBase(BigDecimal yliaoMaxBase) {
         this.yliaoMaxBase = yliaoMaxBase;
     }
 
@@ -346,7 +346,7 @@ public class SocialCitydata {
      *
      * @return yliao_min_base - 医疗保险最低基数
      */
-    public Integer getYliaoMinBase() {
+    public BigDecimal getYliaoMinBase() {
         return yliaoMinBase;
     }
 
@@ -355,7 +355,7 @@ public class SocialCitydata {
      *
      * @param yliaoMinBase 医疗保险最低基数
      */
-    public void setYliaoMinBase(Integer yliaoMinBase) {
+    public void setYliaoMinBase(BigDecimal yliaoMinBase) {
         this.yliaoMinBase = yliaoMinBase;
     }
 
@@ -400,7 +400,7 @@ public class SocialCitydata {
      *
      * @return sye_max_base - 失业保险最高基数
      */
-    public Integer getSyeMaxBase() {
+    public BigDecimal getSyeMaxBase() {
         return syeMaxBase;
     }
 
@@ -409,7 +409,7 @@ public class SocialCitydata {
      *
      * @param syeMaxBase 失业保险最高基数
      */
-    public void setSyeMaxBase(Integer syeMaxBase) {
+    public void setSyeMaxBase(BigDecimal syeMaxBase) {
         this.syeMaxBase = syeMaxBase;
     }
 
@@ -418,7 +418,7 @@ public class SocialCitydata {
      *
      * @return sye_min_base - 失业保险最低基数
      */
-    public Integer getSyeMinBase() {
+    public BigDecimal getSyeMinBase() {
         return syeMinBase;
     }
 
@@ -427,7 +427,7 @@ public class SocialCitydata {
      *
      * @param syeMinBase 失业保险最低基数
      */
-    public void setSyeMinBase(Integer syeMinBase) {
+    public void setSyeMinBase(BigDecimal syeMinBase) {
         this.syeMinBase = syeMinBase;
     }
 
@@ -472,7 +472,7 @@ public class SocialCitydata {
      *
      * @return gs_max_base - 工伤保险最高基数
      */
-    public Integer getGsMaxBase() {
+    public BigDecimal getGsMaxBase() {
         return gsMaxBase;
     }
 
@@ -481,7 +481,7 @@ public class SocialCitydata {
      *
      * @param gsMaxBase 工伤保险最高基数
      */
-    public void setGsMaxBase(Integer gsMaxBase) {
+    public void setGsMaxBase(BigDecimal gsMaxBase) {
         this.gsMaxBase = gsMaxBase;
     }
 
@@ -490,7 +490,7 @@ public class SocialCitydata {
      *
      * @return gs_min_base - 工伤保险最低基数
      */
-    public Integer getGsMinBase() {
+    public BigDecimal getGsMinBase() {
         return gsMinBase;
     }
 
@@ -499,7 +499,7 @@ public class SocialCitydata {
      *
      * @param gsMinBase 工伤保险最低基数
      */
-    public void setGsMinBase(Integer gsMinBase) {
+    public void setGsMinBase(BigDecimal gsMinBase) {
         this.gsMinBase = gsMinBase;
     }
 
@@ -544,7 +544,7 @@ public class SocialCitydata {
      *
      * @return syu_max_base - 生育保险最高基数
      */
-    public Integer getSyuMaxBase() {
+    public BigDecimal getSyuMaxBase() {
         return syuMaxBase;
     }
 
@@ -553,7 +553,7 @@ public class SocialCitydata {
      *
      * @param syuMaxBase 生育保险最高基数
      */
-    public void setSyuMaxBase(Integer syuMaxBase) {
+    public void setSyuMaxBase(BigDecimal syuMaxBase) {
         this.syuMaxBase = syuMaxBase;
     }
 
@@ -562,7 +562,7 @@ public class SocialCitydata {
      *
      * @return syu_min_base - 生育保险最低基数
      */
-    public Integer getSyuMinBase() {
+    public BigDecimal getSyuMinBase() {
         return syuMinBase;
     }
 
@@ -571,7 +571,7 @@ public class SocialCitydata {
      *
      * @param syuMinBase 生育保险最低基数
      */
-    public void setSyuMinBase(Integer syuMinBase) {
+    public void setSyuMinBase(BigDecimal syuMinBase) {
         this.syuMinBase = syuMinBase;
     }
 
@@ -616,7 +616,7 @@ public class SocialCitydata {
      *
      * @return zf_max_base - 住房公积金最高基数
      */
-    public Integer getZfMaxBase() {
+    public BigDecimal getZfMaxBase() {
         return zfMaxBase;
     }
 
@@ -625,7 +625,7 @@ public class SocialCitydata {
      *
      * @param zfMaxBase 住房公积金最高基数
      */
-    public void setZfMaxBase(Integer zfMaxBase) {
+    public void setZfMaxBase(BigDecimal zfMaxBase) {
         this.zfMaxBase = zfMaxBase;
     }
 
@@ -634,7 +634,7 @@ public class SocialCitydata {
      *
      * @return zf_min_base - 住房公积金最低基数
      */
-    public Integer getZfMinBase() {
+    public BigDecimal getZfMinBase() {
         return zfMinBase;
     }
 
@@ -643,7 +643,7 @@ public class SocialCitydata {
      *
      * @param zfMinBase 住房公积金最低基数
      */
-    public void setZfMinBase(Integer zfMinBase) {
+    public void setZfMinBase(BigDecimal zfMinBase) {
         this.zfMinBase = zfMinBase;
     }
